Develop

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -4,7 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
@@ -20,6 +23,11 @@ public class UserService {
     @Qualifier("userDbStorage")
     private final UserStorage inMemoryUserStorage;
 
+    public void deleteUser(Integer id) {
+        log.info("Получен DELETE User");
+        inMemoryUserStorage.deleteUser(id);
+    }
+
     public User saveUser(User user) {
         log.info("Получен POST User");
         return inMemoryUserStorage.postUser(user);
@@ -27,7 +35,7 @@ public class UserService {
 
     public User changeUser(User user) {
         if (inMemoryUserStorage.putUser(user) == null) {
-            return null;
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
         }
         log.info("Получен PUT User");
         return inMemoryUserStorage.putUser(user);
@@ -39,40 +47,44 @@ public class UserService {
 
     public User getUser(Integer id) {
         if (inMemoryUserStorage.getUserId(id) == null) {
-            return null;
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
         }
         return inMemoryUserStorage.getUserId(id);
     }
 
-    public Integer changeFriend(Integer id, Integer friendId) {
-        if (getUser(id) == null) {
-            return id;
+    public void changeFriend(Integer id, Integer friendId) {
+        if (getUser(id) == null || getUser(friendId) == null) {
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
         }
-        if (getUser(friendId) == null) {
-            return friendId;
-        }
-
         inMemoryUserStorage.addFriendId(id, friendId);
         log.info("Получен друг PUT User");
-        return null;
     }
 
-    public Integer deleteFriend(Integer id, Integer friendId) {
-        if (getUser(id) == null) {
-            return id;
-        }
-        if (getUser(friendId) == null) {
-            return friendId;
+    public void deleteFriend(Integer id, Integer friendId) {
+        if (getUser(id) == null || getUser(friendId) == null) {
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
         }
         inMemoryUserStorage.deleteFriendId(id, friendId);
-        return null;
     }
 
     public List<User> getUserFriend(Integer id) {
+        if (getUser(id) == null) {
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
+        }
         return inMemoryUserStorage.getUserIdFriend(id);
     }
 
     public List<User> getCommonFriend(Integer id, Integer otherId) {
+        if (getUser(id) == null || getUser(otherId) == null) {
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
+        }
         return inMemoryUserStorage.getUsersCommonFriends(id, otherId);
+    }
+
+    public List<Film> getUserRecomment(Integer id) {
+        if (getUser(id) == null) {
+            throw new NotFoundException(HttpStatus.NOT_FOUND, "user id not found");
+        }
+        return inMemoryUserStorage.getUserIdRecomment(id);
     }
 }
